# indicate_data_exchange_server/models/aggregated_quality_indicator_value.py
# coding: utf-8

"""
    INDICATE Data Exchange API

    API for exchanging aggregated quality indicator results between INDICATE data providers and the INDICATE hub.  The API has two operations: one for uploading aggregated quality indicator results from a data provider to the hub and one for querying aggregated quality indicator results across all providers. The second operation is intended for use by data providers as well as authorized 3rd-party users. 

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import datetime
from pydantic import BaseModel, ConfigDict, Field, StrictFloat, StrictInt
from typing import Any, ClassVar, Dict, List, Union
from typing_extensions import Annotated
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class AggregatedQualityIndicatorValue(BaseModel):
    """
    AggregatedQualityIndicatorValue
    """ # noqa: E501
    aggregation_period_start: datetime = Field(description="The start point in time of the period to which the aggregated quality indicator value pertains.  The length or end point of the period is either encoded in an adjacent property or evident from the context. ")
    indicator_id: Annotated[int, Field(strict=True, ge=2000000101)] = Field(description="The numeric ID of the OMOP Concept assigned by the INDICATE project for identifying the particular quality indicator. ")
    average_value: Union[StrictFloat, StrictInt] = Field(description="The value of the quality indicator for the indicated period.  The value is a non-negative real number. The value 0 indicates no conformance at all for the period in question while positive number, typically up to 1, indicate some or full conformance. ")
    observation_count: Annotated[int, Field(strict=True, ge=1)] = Field(description="The number of quality indicator results the data provider used to produce the aggregated (average) value. ")
    __properties: ClassVar[List[str]] = ["aggregation_period_start", "indicator_id", "average_value", "observation_count"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of AggregatedQualityIndicatorValue from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of AggregatedQualityIndicatorValue from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "aggregation_period_start": obj.get("aggregation_period_start"),
            "indicator_id": obj.get("indicator_id"),
            "average_value": obj.get("average_value"),
            "observation_count": obj.get("observation_count")
        })
        return _obj